# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrEnvelope:
    properties:
      error:
        type: string
    type: object
  handler.MsgEnvelope:
    properties:
      message:
        type: string
    type: object
...
  model.BookBorrow:
    properties:
      book_id:
        type: integer
      borrow_date:
        type: string
      borrow_id:
        minimum: 1
        type: integer
      return_date:
        type: string
      user_id:
        type: integer
      version:
        type: integer
    required:
    - borrow_id
    type: object
  model.BookBorrowDebtorRp:
    properties:
      book_id:
        type: integer
      book_name:
        type: string
      borrow_date:
        type: string
      borrow_id:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  model.BookBorrowMetricRp:
    properties:
      book_amount:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  model.CreateBookBorrowRq:
    properties:
      book_id:
        minimum: 1
        type: integer
      borrow_date:
        type: string
      user_id:
        minimum: 1
        type: integer
    required:
    - book_id
    - borrow_date
    - user_id
    type: object
...
  model.UpdateBookBorrowRq:
    properties:
      book_id:
        type: integer
      borrow_date:
        type: string
      return_date:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Onelab project
  title: Onelab API
  version: "1.0"
paths:
...
  /book/borrow:
    post:
      consumes:
      - application/json
      description: Create a new book borrow with the input payload
      parameters:
      - description: Book borrow information time:2020-04-17T18:25:43.511Z
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookBorrowRq'
      produces:
      - application/json
      responses:
        "201":
          description: record created
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      summary: Create a new book borrow
      tags:
      - book_borrow
  /book/borrow/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: BookBorrow ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: record deleted
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: record is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Delete book borrow record by id
      tags:
      - book_borrow
    get:
      consumes:
      - application/json
      description: Get borrow record by id in query param
      parameters:
      - description: BookBorrow ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/model.BookBorrow'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: record is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Get Borrow record by id
      tags:
      - book_borrow
    patch:
      consumes:
      - application/json
      parameters:
      - description: BookBorrow ID
        in: path
        name: id
        required: true
        type: integer
      - description: record update information
        in: body
        name: input
        schema:
          $ref: '#/definitions/model.UpdateBookBorrowRq'
      produces:
      - application/json
      responses:
        "200":
          description: record updated
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: record is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "409":
          description: edit conflict
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Update book borrow by id
      tags:
      - book_borrow
  /book/borrow/debtor/list:
    get:
      consumes:
      - application/json
      description: List borrow debtor record
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            items:
              $ref: '#/definitions/model.BookBorrowDebtorRp'
            type: array
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: record is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: List Borrow debtors
      tags:
      - book_borrow
  /book/borrow/metric/list/{id}:
    get:
      consumes:
      - application/json
      description: List borrow debtor record
      parameters:
      - description: Month
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BookBorrowMetricRp'
            type: array
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: record is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: List Borrow metric
      tags:
      - book_borrow
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"